Tweaked the blog.
I've just about got the blog engine finished up.  Pretty happy with
how the back end is working, still some UI work yet to do.

I think at this point I'm going to introduce a new component,
content management.  The blog engine could serve perfectly as an
engine to manage the content on any of the pages of the website.